requirements.txt
# ETL & Data Processing
akshare>=1.12.0
tushare>=1.2.89
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Environment
python-dotenv>=1.0.0

# Scheduling (optional, for automated updates)
schedule>=1.2.0

# Data validation
pydantic>=2.0.0
